# src/main/java/organizationmanagement/dto/DepartmentCreateDTO.java
package organizationmanagement.dto;

public class DepartmentCreateDTO {
    private String name;
    private Long organizationId;

    
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }

    public Long getOrganizationId() { return organizationId; }
    public void setOrganizationId(Long organizationId) { this.organizationId = organizationId; }
}