# src/main/java/organizationmanagement/dto/TeamCreateDTO.java
package organizationmanagement.dto;

public class TeamCreateDTO {
    private String name;
    private Long departmentId;

    // Getters and Setters
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }

    public Long getDepartmentId() { return departmentId; }
    public void setDepartmentId(Long departmentId) { this.departmentId = departmentId; }
}